CoinDiggers Develop

+++ fermat-android-core/src/main/java/com/bitdubai/android_core/app/NotificationService.java
@@ -154,7 +154,7 @@ public class NotificationService extends Service {
     public int notificateProgress(FermatBundle bundle) {
         try {
             int progress = (int) bundle.getSerializable(Broadcaster.PROGRESS_BAR);
-            int publishId = (bundle.contains(Broadcaster.PUBLISH_ID)) ? (int) bundle.getInt(Broadcaster.PUBLISH_ID):0;
+            int publishId = (bundle.contains(Broadcaster.PUBLISH_ID)) ? bundle.getInt(Broadcaster.PUBLISH_ID) :0;
             String progressText = (bundle.contains(Broadcaster.PROGRESS_BAR_TEXT)) ? bundle.getString(Broadcaster.PROGRESS_BAR_TEXT):null;
 
             mNotifyManager = (NotificationManager)
